Refactor transaction decoding to server-side with improved error handling
